+++ prince-archiver/compose.override.yml
@@ -34,19 +34,19 @@ services:
         condition: service_completed_successfully
     <<: *code-and-data-volumes
 
-  event-ingester:
-    environment:
-      SRC_SYSTEMS_DELETE: '["prince"]'
-    <<: *code-and-data-volumes
-
   state-manager:
     environment:
       RABBITMQ_DSN: amqp://guest:guest@rabbitmq:5672
     <<: *code-volumes
 
-  upload-worker:
+  exporter:
     environment:
       <<: *aws-env-variables
+    <<: *code-and-data-volumes
+
+  purger:
+    environment:
+      SRC_SYSTEMS_DELETE: '["prince"]'
     <<: *code-and-data-volumes
 
   api:
