+++ prince-archiver/compose.prod.yml
@@ -13,19 +13,6 @@ x-log-labels: &log-labels
 
 
 services:
-  event-ingester:
-    environment:
-      SENTRY_DSN: ${SENTRY_DSN}
-    volumes:
-      - cifs_mount:/data/prince
-    labels: *log-labels
-    deploy:
-      restart_policy:
-        condition: on-failure
-    env_file:
-      - .env
-      - ../.env
-
   state-manager:
     environment:
       <<: *aws-env-variables
@@ -39,12 +26,25 @@ services:
       - .env
       - ../.env
 
-  upload-worker:
+  exporter:
     environment:
       <<: *aws-env-variables
       SENTRY_DSN: ${SENTRY_DSN}
     volumes:
       - cifs_mount:/data/prince:ro
+    labels: *log-labels
+    deploy:
+      restart_policy:
+        condition: on-failure
+    env_file:
+      - .env
+      - ../.env
+
+  purger:
+    environment:
+      SENTRY_DSN: ${SENTRY_DSN}
+    volumes:
+      - cifs_mount:/data/prince
     labels: *log-labels
     deploy:
       restart_policy:
