+++ prince-archiver/compose.yml
@@ -9,22 +9,10 @@ x-redis-env-variable: &redis-env-variable
 
 
 services:
-  event-ingester:
+  state-manager:
     image: mycostreams/prince-archiver:${TAG:-latest}
     build:
       context: .
-    command: ["arq", "prince_archiver.entrypoints.event_ingester.WorkerSettings"]
-    environment: 
-      <<: *redis-env-variable
-      SRC_DIR: /data
-    depends_on:
-      prestart:
-        condition: service_completed_successfully
-    networks:
-      - default
-
-  state-manager:
-    image: mycostreams/prince-archiver:${TAG:-latest}
     command: ["arq", "prince_archiver.entrypoints.state_manager.WorkerSettings"]
     environment:
       <<: [*postgres-env-variable, *redis-env-variable]
@@ -34,9 +22,21 @@ services:
     networks:
       - default
 
-  upload-worker:
+  exporter:
     image: mycostreams/prince-archiver:${TAG:-latest}
-    command: ["arq", "prince_archiver.entrypoints.upload_worker.WorkerSettings"]
+    command: ["arq", "prince_archiver.entrypoints.exporter.WorkerSettings"]
+    environment: 
+      <<: *redis-env-variable
+      SRC_DIR: /data
+    depends_on:
+      prestart:
+        condition: service_completed_successfully
+    networks:
+      - default
+
+  purger:
+    image: mycostreams/prince-archiver:${TAG:-latest}
+    command: ["arq", "prince_archiver.entrypoints.purger.WorkerSettings"]
     environment: 
       <<: *redis-env-variable
       SRC_DIR: /data
