+++ prince-archiver/prince_archiver/entrypoints/purger/settings.py
